# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/song": {
            "post": {
                "description": "Create a new song in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "Create a new song",
                "parameters": [
                    {
                        "description": "Song data",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Create"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Song"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/song/{id}": {
            "put": {
                "description": "Update an existing song in the system by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "Update a song by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated song data",
                        "name": "song",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Song"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Song"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a song from the system by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "Delete a song by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/songs/search": {
            "post": {
                "description": "Search for songs based on search criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "Search for songs",
                "parameters": [
                    {
                        "description": "Search parameters",
                        "name": "search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Search"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Entity"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "get": {
                "description": "Get a song by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "Get a song by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Entity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/songs/{id}/text": {
            "post": {
                "description": "Get the text of a song by its ID with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "songs"
                ],
                "summary": "Get text of a song by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pagination parameters",
                        "name": "pagination",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Pagination"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Create": {
            "type": "object",
            "required": [
                "group",
                "name"
            ],
            "properties": {
                "group": {
                    "description": "Группа, к которой принадлежит песня",
                    "type": "string"
                },
                "name": {
                    "description": "Название песни",
                    "type": "string"
                }
            }
        },
        "Entity": {
            "type": "object",
            "required": [
                "group",
                "name"
            ],
            "properties": {
                "group": {
                    "description": "Группа, к которой принадлежит песня",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "description": "Ссылка на песню",
                    "type": "string"
                },
                "name": {
                    "description": "Название песни",
                    "type": "string"
                },
                "releaseDate": {
                    "description": "Дата релиза в формате yyyy.dd.mm",
                    "type": "string",
                    "example": "2023.15.01"
                },
                "text": {
                    "description": "Текст песни *ПРИПЕВЫ выделять -Начало припева- .... -Конец припева-",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "Pagination": {
            "type": "object",
            "required": [
                "page",
                "page_size"
            ],
            "properties": {
                "page": {
                    "description": "Номер страницы",
                    "type": "integer"
                },
                "page_size": {
                    "description": "Количество элементов на странице",
                    "type": "integer"
                }
            }
        },
        "Search": {
            "type": "object",
            "required": [
                "page",
                "page_size"
            ],
            "properties": {
                "group": {
                    "description": "Группа для поиска",
                    "type": "string"
                },
                "link": {
                    "description": "Ссылка на песню",
                    "type": "string"
                },
                "name": {
                    "description": "Название песни для поиска",
                    "type": "string"
                },
                "page": {
                    "description": "Номер страницы",
                    "type": "integer"
                },
                "page_size": {
                    "description": "Количество элементов на странице",
                    "type": "integer"
                },
                "releaseDate": {
                    "description": "Дата релиза в формате yyyy.dd.mm",
                    "type": "string",
                    "example": "2023.15.01"
                },
                "text": {
                    "description": "Текст песни *ПРИПЕВЫ выделять -Начало припева- .... -Конец припева-",
                    "type": "string",
                    "example": ""
                }
            }
        },
        "Song": {
            "type": "object",
            "required": [
                "group",
                "name"
            ],
            "properties": {
                "group": {
                    "description": "Группа, к которой принадлежит песня",
                    "type": "string"
                },
                "link": {
                    "description": "Ссылка на песню",
                    "type": "string"
                },
                "name": {
                    "description": "Название песни",
                    "type": "string"
                },
                "releaseDate": {
                    "description": "Дата релиза в формате yyyy.dd.mm",
                    "type": "string",
                    "example": "2023.15.01"
                },
                "text": {
                    "description": "Текст песни *ПРИПЕВЫ выделять -Начало припева- .... -Конец припева-",
                    "type": "string",
                    "example": ""
                }
            }
        }
    }
}